Refactor Lexer
